+++ conftest.py
@@ -32,9 +32,11 @@ class TestCase:
         return self
 
     def run(self):
-        assert self._registry.is_healthy
         self._registry.enabled_checks = {}
         handlers = self._registry.bind()
+        assert (
+            self._registry.is_healthy
+        ), f"Settings has errors: {self._registry._config.errors.as_text()}"
         return list(handlers[self.TEST_TAG]())
 
     def models(self, *models):