+++ src/extra_checks/forms.py
@@ -1,3 +1,4 @@
+import importlib
 import typing
 
 import django.core.checks
@@ -165,11 +166,35 @@ class BaseCheckForm(forms.Form):
         choices=[(c, c) for c in ["DEBUG", "INFO", "WARNING", "ERROR", "CRITICAL"]],
         required=False,
     )
+    ignore_types = ListField(forms.CharField(), required=False)
 
     def clean_level(self) -> typing.Optional[int]:
         if self.cleaned_data["level"]:
             return getattr(django.core.checks, self.cleaned_data["level"])
         return None
+
+    def clean_ignore_types(self) -> set:
+        value = self.cleaned_data["ignore_types"]
+        if not value:
+            return value
+        result = []
+        for import_path in value:
+            try:
+                path, entry = import_path.rsplit(".", 1)
+                result.append(getattr(importlib.import_module(path), entry))
+            except (ImportError, ValueError, AttributeError):
+                raise forms.ValidationError(
+                    f"ignore_types contains entry that can't be imported: '{import_path}'."
+                )
+        return result
+
+    def clean(self) -> typing.Dict[str, typing.Any]:
+        if (
+            "ignore_types" in self.cleaned_data
+            and not self.cleaned_data["ignore_types"]
+        ):
+            del self.cleaned_data["ignore_types"]
+        return self.cleaned_data
 
 
 class AttrsForm(BaseCheckForm):