+++ tests/test_config.py
@@ -1,12 +1,14 @@
 from django.core import checks
 
-from extra_checks import CheckId, ignore_checks
+from extra_checks import CheckId
 from extra_checks.checks import (
     CheckFieldFileUploadTo,
     CheckModelAttribute,
     CheckModelMetaAttribute,
 )
 from extra_checks.forms import ConfigForm
+from extra_checks.registry import ChecksConfig
+from tests.example.models import Article, Author
 
 
 def test_config_form():
@@ -121,10 +123,41 @@ def test_config_include_apps():
     assert form.cleaned_data == {"include_apps": ["tests.example"], "checks": {}}
 
 
-def test_ignore_checks(monkeypatch):
-    ignore: dict = {}
-    monkeypatch.setattr("extra_checks._IGNORED", ignore)
+def test_ignore_checks(registry):
     obj = object()
-    res = ignore_checks("X010", "model-attribute", CheckId.X050)(obj)
+    res = registry.ignore_checks("X010", "model-attribute", CheckId.X050)(obj)
     assert res == obj
-    assert ignore == {obj: {"X010", "model-attribute", CheckId.X050}}
+    assert registry.ignored_checks == {obj: {"X010", "model-attribute", CheckId.X050}}
+
+
+def test_config_build_ignored():
+    ignored, errors = ChecksConfig._build_ignored(
+        {
+            Article: {"X010", "random-test-name"},
+            Author: {"model-meta-attribute", CheckId.X050},
+        }
+    )
+    assert errors == [
+        "Unknown check (random-test-name) provided to the 'ignore_checks'."
+    ]
+    assert ignored == {
+        CheckId.X010: {Article},
+        CheckId.X011: {Author},
+        CheckId.X050: {Author},
+    }
+
+
+def test_config_form_ignore_types():
+    form = ConfigForm(
+        data={
+            "checks": [
+                {
+                    "id": CheckFieldFileUploadTo.Id.value,
+                    "ignore_types": ["django.db.models.ImageField"],
+                }
+            ]
+        }
+    )
+    assert form.is_valid(
+        {CheckFieldFileUploadTo.Id: CheckFieldFileUploadTo.settings_form_class}
+    )